RT#70346: delete_subfield() no longer clears entire field if arguments are badly formatted
Prior to this patch, $field->delete_subfield('a') used to clear
the entire field.  Now, $field->delete_subfield('a') will DWIM,
making it equivalent to $field->delete_subfields('a').

delete_subfield() will now croak if it's not called with
either a single scalar or a hash containing at least
one of the following keys: code, pos, or match.A

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>