use XML::LibXML::Reader to pull in records
This is more correct than the previous approach of
setting $/ to 'record>', since it can handle self-closing
empty record elements.

In order to provide a speed boost, MARC::File::XML::decode
now checks to see if it's dealing with an XML string or
an XML::LibXML::Element object containing a MARCXML
tree.  That may well be useful outside of MARC::Batch processing.

There are two issues with this patch as it stands, however:

[1] It makes processing a batch 15%.  That may be an ecceptable
    tradeoff, particularly if an option is added to reenable
    the looser but faster record slurping.
[2] It introduces a memory leak; perhaps the XML::LibXML::Element
    returned by _next is not getting freed?

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>