use G0 versions of characters in character DB
The file codetables.xml distributed by the Library
of Congress provides the G0 MARC8 codepoint for
"basic" character sets and the G1 MARC8 codepoint
for "extended" ones.  This patch normalizes everything
to G0, with the exception of some C0 and C1 characters.

This is part of the fix for RT#63271.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>