recognize G0 and G1 characters properly
When converting from MARC8 to UTF8, MARC::Charset now
properly recognizes if a (single-byte) MARC8 character falls
in G0 or G1.

This is part of the fix for RT#63271 (converting characters
in the Extended Cyrillic character set), but should also
fix similar issues with converting characters in the extended
Arabic set.

This commit also means that all MARC8 character sets that support
both G0 and G1 wll be properly converted, regardless of whether
they're currently set as the G0 or G1 character set.  For example,
it is now possible to convert Extended Latin as G0 or Basic Latin
as G1.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>