do not process external entities by default
Allowing external XML entities to be processed when parsing a MARCXML
document can result in unwanted disclosure of the contents of files on
the server, which could be a security vulnernability for applications
that embed MARC::File::XML.

This patch will make MARC::File::XML die upon parsing a record that
declares an external entity and tries to use it.  If, for some reason,
an application needs to process MARCXML records that contain external
entities, set_parser() can be used to force the use of an XML::LibXML
parser that is configured to process external entities.

The issue was reported by John Lightsey.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>